requirements.txt
Django>=4.0
Pillow>=9.0
requests>=2.28
fastapi>=0.95
uvicorn>=0.21
firebase-admin>=6.1
pydantic>=1.10
drf-spectacular>=0.29.0
djangorestframework>=3.15.2
django-cors-headers>=4.4.0
google-generativeai>=0.8.5
